cruce_stock/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
PyYAML>=6.0
chardet>=5.0.0
streamlit>=1.32.0
